Modified function declarations: fixed function and return types

+++ header.h
@@ -4,6 +4,6 @@
 //function declararations here
 int total_trip_days();
 double round_trip_airfare();
-double private_vehicle_cost(int miles_driven);
-int departure_arrival_times(double departure_time, double arrival_time);
+double private_vehicle_cost(double miles_driven);
+double departure_arrival_times(double departure_time, double arrival_time);
 #endif
